third_party/blink/web_tests/platform/generic/external/wpt/orientation-event/orientation/horizontal-surface-manual.https-expected.txt

This is a testharness.js-based test.
FAIL X angle assert_approx_equals: expected a number but got a "undefined"
FAIL Y angle assert_approx_equals: expected a number but got a "undefined"
Harness: the test ran to completion.
